requirements.txt
streamlit
pdfplumber
python-docx
pytest